Impleement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -65,7 +65,14 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        // if(\Gate::allows('update-question',$question))
+        if(\Gate::denies('update-question',$question))
+        {
+            abort(403,"access denied");
+
+        }
         return view("questions.edit",compact('question'));
+
     }
 
     /**
@@ -77,8 +84,13 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        $question->update($request->only('title','body'));
-         return redirect()->route('questions.index')->with('success',"Your question has been updated.");
+        if(\Gate::denies('update-question',$question))
+        {
+            abort(403,"access denied");
+
+        }
+            $question->update($request->only('title','body'));
+            return redirect()->route('questions.index')->with('success',"Your question has been updated.");
     }
 
     /**
@@ -89,6 +101,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question))
+        {
+            abort(403,"access denied");
+
+        }
         $question->delete();
         return redirect('/questions')->with('success',"Question has been deleted");
     }
